fixes to lab6

+++ lab6/filtr_doln.m
@@ -2,7 +2,7 @@ clearvars; close all; clc;
 image = imread('plansza.bmp');
 
 filtrAvg = fspecial('average', 3);
-conv = conv2(image, filtrAvg, 'same');
+conv = conv2(double(image), filtrAvg, 'same');
 conv = uint8(conv);
 
 subplot(3,3,1); imshow(image); title('oryginal');
@@ -13,7 +13,7 @@ subplot(3,3,3); imshow(imabsdiff(conv, image), []); title('imabsdiff');
 M = [1 2 1; 2 4 2; 1 2 1];
 M = M / sum(sum(M));
 
-convM = conv2(image, M, 'same');
+convM = conv2(double(image), M, 'same');
 convM = uint8(convM);
 
 subplot(3,3,4); imshow(image); title('oryginal');
@@ -22,7 +22,7 @@ subplot(3,3,6); imshow(imabsdiff(convM, image), []); title('imabsdiff');
 
 %gaussian
 filtrG = fspecial('gaussian', 5, 0.5);
-convG = uint8(conv2(image, filtrG, 'same'));
+convG = uint8(conv2(double(image), filtrG, 'same'));
 
 subplot(3,3,7); imshow(image); title('oryginal');
 subplot(3,3,8); imshow(convG); title('gaussian');
